Move inflate rendering to separate namespace
In order to support arbitrary (ClojureScript data) keys, we'll need to use
`cljs.reader` to deserialize keys from string markup. This namespace is
currently not efficiently tree-shaked, so we want to keep it out of our code
base entirely if we're not using server-rendering, as bundle sizes will increase
a little.