Do not attach the update hook on a component's initial render
Wow, this was a doozy. Because dumdom converts ClojureScript data to JavaScript
objects on every call to dumdom.core/render, Snabbdom will sometimes call the
update hook on DOM elements even when there is no relevant change in the data
rendered by dumdom. Snabbdom operates on identity semantics, and will fail to
notice that two vdom nodes are equal per ClojureScript's value semantics.

With this change, the update hook is not added to the vdom node at all until we
know it is relevant to do so. When it is added, it is wrapped with some state
that ensures it is only ever called once.