Make Snabbdom fully component-aware via keys
Snabbdom, dumdom'd underlying VDOM rendering engine does not have a concept of
components. A component is a dumdom abstraction. This commit fixes some nasty
issues with component lifecycles that were not properly handled between dumdom
and snabbdom.

Because snabbdom does not support components, a dumdom component in practice is
it's root _element_. This means that the root element of a component will
receive life-cycle events from snabbdom, and dumdom will relay these to the
component life-cycle callbacks. However, the component and element essentially
being the same thing caused some unintended side-effects.

Because components are their root element, snabbdom is not aware of components,
and is unable to tell that two DOM trees originate from different components if
they are similar enough:

```clj
(q/defcomponent Child1 [props]
  [:div "Child 1"])

(q/defcomponent Child2 [props]
  [:div "Child 2"])

(def data {:one? true})

(q/render [:div (if (:one? data)
                  (Child1 data)
                  (Child2 data))] el)
```

In this example, dumdom would have snabbdom render one of the following:

```html
<div><div>Child 1</div></div>
```

Or:

```html
<div><div>Child 2</div></div>
```

The quickest way to achieve this change is to just change the text in the
elements, so snabbdom would not need to destroy or create any elements. Thus it
would not emit any life-cycle events. Dumdom, in turn, would not call on
`:on-unmount` for `Child1` and then `:on-mount` for `Child2`, as you would
expect.

This problem could get really gnarly if a component rendered another component
in its root position, as now there are two components and one element that are
essentially the same thing. If the parent conditionally renders one of several
components in its root position, you would be susceptible to some very strange
behavior unless you meticulously used `:keyfn` on everything.

Relying on `:keyfn` is not a good solution, as it becomes up to dumdom's users
whether or not it would behave deterministically. Clearly not ideal. So instead,
this commit changes how dumdom distributes keys.

Dumdom now gives every element created by a component a key, regardless of
whether the user provided one. The key from the component (blank or not) will
now be embellished with the component that created it, potentially nesting
several levels. With this change, dumdom exposes components to snabbdom via
element keys, and life-cycle events always dispatch as expected in dumdom.