Use a global binding for the top-level event handler
As it turns out, Dumdom can't properly honor the contract promised by the map
passed to the render function, because it creates a long-lived stateful mess...
So let's not invite people into that trap.

Disclaimer: It will seem to work if you never change the `:handle-event`
function, so it's not entirely broken.